ref(select): move ApiV3 and ApiV4 directives in core-select
