+++ packages/ng/core-select/api/public-api.ts
@@ -1,0 +1,3 @@
+export * from './api-v3.directive';
+export * from './api-v4.directive';
+export { ALuCoreSelectApiDirective } from './api.directive';
